CI: Add build-tests for various compilers on Windows and Ubuntu
The build setup for this repository is incredibly simple, but
unfortunately even recent changes to the CMake have made various
platforms and compiler combinations unbuildable.  Add a basic CI that at
the very least _builds_ the project to demonstrate all these problems,
which should all go away when other PRs are merged.